+++ src/Recipient/AndroidDeviceRecipient.php
@@ -15,7 +15,7 @@ final class AndroidDeviceRecipient implements RecipientInterface
     /**
      * @param string $token
      */
-    public function __construct($token)
+    public function __construct(string $token)
     {
         $this->token = $token;
     }